+++ src/uk/ac/cam/cl/groupprojectdelta/galtonboards/graphics/UserInterface.java
@@ -18,33 +18,28 @@ import static org.lwjgl.system.MemoryUtil.NULL;
 
 public class UserInterface {
 
-  private final Window windowBoards;
-  private final Window windowControls;
-  private boolean running = false;
+  private final WindowBoards windowBoards;
 
-  UserInterface(Window windowBoards, Window windowControls) {
+  UserInterface(WindowBoards windowBoards) {
     this.windowBoards = windowBoards;
-    this.windowControls = windowControls;
   }
 
   /**
    * Initialize OpenGL and all the windows
    * Then, run the main loop
-   * @param wb : WindowBoards
-   * @param wc : WindowControls
    */
-  public void start(WindowBoards wb, WindowControls wc) {
-    wc.addComponent(makeButton(64, 32, 32, 0xF40A,
-            (MouseClickEventListener) event -> wb.getSimulation().run()));
-    wc.addComponent(makeButton(64, 128, 32, 0xF3E4,
-            (MouseClickEventListener) event -> wb.getSimulation().pause()));
-    wc.addComponent(makeButton(64, 224, 32, 0xF4DB,
-            (MouseClickEventListener) event -> wb.getSimulation().stop()));
+  public void start() {
+    windowBoards.addComponent(makeButton(64, 32, 32, 0xF40A,
+            (MouseClickEventListener) event -> windowBoards.getSimulation().run()));
+    windowBoards.addComponent(makeButton(64, 128, 32, 0xF3E4,
+            (MouseClickEventListener) event -> windowBoards.getSimulation().pause()));
+    windowBoards.addComponent(makeButton(64, 224, 32, 0xF4DB,
+            (MouseClickEventListener) event -> windowBoards.getSimulation().stop()));
 
     Button resetViewButton = new Button("Reset view", 80, 128, 160, 32);
     resetViewButton.getStyle().setBorder(new SimpleLineBorder(ColorConstants.black(), 1));
-    resetViewButton.getListenerMap().addListener(MouseClickEvent.class, e -> wb.getCamera().Reset());
-    wc.addComponent(resetViewButton);
+    resetViewButton.getListenerMap().addListener(MouseClickEvent.class, e -> windowBoards.getCamera().Reset());
+    windowBoards.addComponent(resetViewButton);
 
     System.setProperty("joml.nounsafe", Boolean.TRUE.toString());
     System.setProperty("java.awt.headless", Boolean.TRUE.toString());
@@ -68,35 +63,13 @@ public class UserInterface {
     GL.createCapabilities();
     windowBoards.initialize(windowBoardsID);
 
-    // Initialize window for controls
-    long windowControlsID = glfwCreateWindow(windowControls.getWidth(), windowControls.getHeight(), "Controls", NULL, NULL);
-    if (windowControlsID == NULL) throw new RuntimeException("Failed to create the GLFW window");
-    glfwShowWindow(windowControlsID);
-    glfwMakeContextCurrent(windowControlsID);
-    GL.createCapabilities();
-    windowControls.initialize(windowControlsID);
-
-
     // Main loop
-    running = true;
-    while (running) {
-      if (glfwWindowShouldClose(windowBoardsID) || glfwWindowShouldClose(windowControlsID)) {
-        running = false;
-        break;
-      }
-
-      glfwMakeContextCurrent(windowBoardsID);
-      GL.getCapabilities();
+    while (!glfwWindowShouldClose(windowBoardsID)) {
       windowBoards.loop(windowBoardsID);
-
-      glfwMakeContextCurrent(windowControlsID);
-      GL.getCapabilities();
-      windowControls.loop(windowControlsID);
     }
 
     // Destroy windows
     windowBoards.destroy(windowBoardsID);
-    windowControls.destroy(windowControlsID);
   }
 
   private static Button makeButton(int size, int xPos, int yPos, int iconCode, EventListener<MouseClickEvent> cb) {
