+++ src/uk/ac/cam/cl/groupprojectdelta/galtonboards/graphics/WindowBoards.java
@@ -1,6 +1,17 @@
 package uk.ac.cam.cl.groupprojectdelta.galtonboards.graphics;
 
+import java.util.Vector;
 import org.joml.Matrix4f;
+import org.joml.Vector2f;
+import org.liquidengine.legui.DefaultInitializer;
+import org.liquidengine.legui.animation.AnimatorProvider;
+import org.liquidengine.legui.component.Component;
+import org.liquidengine.legui.component.Frame;
+import org.liquidengine.legui.input.Mouse;
+import org.liquidengine.legui.listener.processor.EventProcessorProvider;
+import org.liquidengine.legui.style.color.ColorConstants;
+import org.liquidengine.legui.system.layout.LayoutManager;
+import org.lwjgl.BufferUtils;
 import uk.ac.cam.cl.groupprojectdelta.galtonboards.UserInput;
 import uk.ac.cam.cl.groupprojectdelta.galtonboards.workspace.Configuration;
 import uk.ac.cam.cl.groupprojectdelta.galtonboards.workspace.Simulation;
@@ -29,18 +40,36 @@ class WindowBoards extends Window {
   private int vertexShaderID;
   private int fragmentShaderID;
   private int vertexBuffer, uvBuffer;
+  private int textureID;
 
   private final String textureFilePath;
   private final String vertexShaderPath;
   private final String fragmentShaderPath;
 
-  final float[] CLEAR_COLOUR = {0.5f, 0.5f, 0.5f, 1};
+  // UI
+  private Frame frame;
+  private DefaultInitializer initializer;
+  private final Vector<Component> components = new Vector<>();
+  private boolean initialized = false;
+
+  static final float[] CLEAR_COLOUR = {0.5f, 0.5f, 0.5f, 1};
 
   public WindowBoards(int width, int height, String vertexShaderPath, String fragmentShaderPath, String textureFilePath) {
     super(width, height);
     this.vertexShaderPath = vertexShaderPath;
     this.fragmentShaderPath = fragmentShaderPath;
     this.textureFilePath = textureFilePath;
+  }
+
+  /**
+   * Add legui component to the window
+   * Can only be called before starting the main loop
+   */
+  void addComponent(Component component) {
+    // todo: this should throw an exception instead of failing silently
+    if (!initialized) {
+      components.add(component);
+    }
   }
 
   Configuration getConfiguration() {
@@ -76,8 +105,34 @@ class WindowBoards extends Window {
     vertexBuffer = glGenBuffers();
     uvBuffer = glGenBuffers();
 
+    glEnable(GL_BLEND);
+    glBlendFunc(GL_SRC_ALPHA, GL_ONE_MINUS_SRC_ALPHA);
+    glDisable(GL_CULL_FACE);
+    glDisable(GL_DEPTH_TEST);
+
+    // enable V-sync
+    glfwSwapInterval(1);
+
     // Load texture
-    loadTexture(textureFilePath);
+    textureID = loadTexture(textureFilePath);
+
+    // Create the UI Frame
+    frame = new Frame(getWidth(), getHeight());
+    frame.getContainer().getStyle().getBackground().setColor(ColorConstants.transparent());
+    frame.getContainer().setFocusable(false);
+    for (Component component : components) {
+      frame.getContainer().add(component);
+    }
+
+    // Create GUI initializer
+    initializer = new DefaultInitializer(window, frame);
+
+    // Initialize renderer
+    initializer.getRenderer().initialize();
+
+    initialized = true;
+
+    glClearColor(CLEAR_COLOUR[0], CLEAR_COLOUR[1], CLEAR_COLOUR[2], CLEAR_COLOUR[3]);
   }
 
   @Override
@@ -102,6 +157,12 @@ class WindowBoards extends Window {
     deltaTime = (currentTime - lastTime);
     userInput.update(deltaTime);
     workspace.update(deltaTime);
+
+    // enable transparency
+    glEnable(GL_BLEND);
+    glBlendFunc(GL_SRC_ALPHA, GL_ONE_MINUS_SRC_ALPHA);
+    // use GL shader
+    glUseProgram(programID);
 
     int[] windowWidth = new int[1];
     int[] windowHeight = new int[1];
@@ -146,11 +207,22 @@ class WindowBoards extends Window {
     glVertexAttribPointer(1, 2, GL_FLOAT, false, 0, 0);
     glEnableVertexAttribArray(1);
 
+    glBindTexture(GL_TEXTURE_2D, textureID);
+
     glDrawArrays(
         GL_TRIANGLES,
         0,
         mesh.length / 3
     );
+
+    initializer.getContext().updateGlfwWindow();
+    initializer.getRenderer().render(frame, initializer.getContext());
+
+    // Process events
+    initializer.getSystemEventProcessor().processEvents(frame, initializer.getContext());
+    EventProcessorProvider.getInstance().processEvents();
+    LayoutManager.getInstance().layout(frame);
+    AnimatorProvider.getAnimator().runAnimations();
 
     glfwSwapBuffers(window);
     glfwPollEvents();
@@ -158,6 +230,9 @@ class WindowBoards extends Window {
 
   @Override
   void destroy(long window) {
+
+    initializer.getRenderer().destroy();
+
     // Detach and delete shaders
     glDetachShader(programID, vertexShaderID);
     glDetachShader(programID, fragmentShaderID);
@@ -167,6 +242,8 @@ class WindowBoards extends Window {
     glDeleteProgram(programID);
     glfwFreeCallbacks(window);
     glfwDestroyWindow(window);
+
+    glfwTerminate();
   }
 
   /**
@@ -200,7 +277,7 @@ class WindowBoards extends Window {
   /**
    * Load a texture
    *
-   * @param texturePath: shader file name
+   * @param texturePath: texture file name
    * @return texture identifier
    */
   private int loadTexture(String texturePath) {
